+++ library/FieldUtils/InputModules/InputNek5000.cpp
@@ -55,28 +55,6 @@ ModuleKey InputNek5000::m_className[1] = {
 };
 
 /**
- * @brief Swap endian ordering of the input variable.
- */
-template <typename T>
-void swap_endian(T &u)
-{
-    union
-    {
-        T u;
-        unsigned char u8[sizeof(T)];
-    } source, dest;
-
-    source.u = u;
-
-    for (size_t k = 0; k < sizeof(T); k++)
-    {
-        dest.u8[k] = source.u8[sizeof(T) - k - 1];
-    }
-
-    u = dest.u;
-}
-
-/**
  * @brief Set up InputNek5000 object.
  *
  */