+++ CHANGELOG.md
@@ -11,6 +11,9 @@ v4.5.0
 - Added in sum factorisation version for pyramid expansions and orthogonal
   expansion in pyramids (!750)
 
+**FieldConvert**:
+- Add input module for Semtex field files (!777)
+
 **Documentation**:
 - Added the developer-guide repository as a submodule (!751)
 
@@ -25,6 +28,8 @@ v4.4.1
 - Fix issue with average fields not working with different polynomial order
   fields (!776)
 - Fix Hdf5 output in FilterFieldConvert (!781)
+- Fixed extreme memory consumption of Interpolator when interpolating from pts
+  to fld or between different meshes (!783)
 
 **FieldConvert:**
 - Fix issue with field ordering in the interppointdatatofld module (!754)